+++ src/main/scala/net/gonzberg/spark/sorting/util/GroupByKeyIterator.scala
@@ -1,4 +1,4 @@
-package net.gonzberg.spark.sorting
+package net.gonzberg.spark.sorting.util
 
 private[sorting] class GroupByKeyIterator[K, V](iter: Iterator[(K, V)])
     extends Iterator[(K, Iterator[V])] {
