+++ src/main/scala/net/gonzberg/spark/sorting/util/SortHelpers.scala
@@ -1,4 +1,4 @@
-package net.gonzberg.spark.sorting
+package net.gonzberg.spark.sorting.util
 
 import org.apache.spark.Partitioner
 import org.apache.spark.rdd.RDD
@@ -7,9 +7,9 @@ import scala.reflect.ClassTag
 
 private[sorting] object SortHelpers {
   def repartitionAndSort[K: Ordering, V: Ordering](
-    rdd: RDD[(K, V)],
-    partitioner: Partitioner
-  ): RDD[(K, V)] = {
+                                                    rdd: RDD[(K, V)],
+                                                    partitioner: Partitioner
+                                                  ): RDD[(K, V)] = {
     rdd
       .mapPartitions(
         partition => {
@@ -26,10 +26,10 @@ private[sorting] object SortHelpers {
   }
 
   def repartitionAndSort[K: Ordering, V: ClassTag, A: Ordering](
-    rdd: RDD[(K, V)],
-    by: V => A,
-    partitioner: Partitioner
-  ): RDD[(K, V)] = {
+                                                                 rdd: RDD[(K, V)],
+                                                                 by: V => A,
+                                                                 partitioner: Partitioner
+                                                               ): RDD[(K, V)] = {
     rdd
       .mapPartitions(
         partition => {
@@ -49,21 +49,22 @@ private[sorting] object SortHelpers {
   }
 
   def modifyResourcePreparationAndOp[R, R1, V, A](
-    prepareResource: R => R1,
-    op: (R1, V) => A
-  ): R => V => A = {
+                                                   prepareResource: R => R1,
+                                                   op: (R1, V) => A
+                                                 ): R => V => A = {
     def newOp(resource: R): V => A = {
       val preparedResource = prepareResource(resource)
       op(preparedResource, _)
     }
+
     newOp
   }
 
   def joinAndApply[K, R, V, A](
-    op: R => V => A
-  )(resourcesIter: Iterator[(K, R)], valuesIter: Iterator[(K, Iterator[V])])(
-    implicit keyOrdering: Ordering[K]
-  ): Iterator[(K, A)] = {
+                                op: R => V => A
+                              )(resourcesIter: Iterator[(K, R)], valuesIter: Iterator[(K, Iterator[V])])(
+                                implicit keyOrdering: Ordering[K]
+                              ): Iterator[(K, A)] = {
     import keyOrdering.mkOrderingOps
 
     val resourceOptionIter = resourcesIter.map(Some(_))
